// src/app/core/services/generated/user-profile/user-profile.service.ts
interface HttpClientOptions {
  headers?: HttpHeaders | Record<string, string | string[]>;
  context?: HttpContext;
  params?:
        | HttpParams
        | Record<string, string | number | boolean | ReadonlyArray<string | number | boolean>>;
  reportProgress?: boolean;
  withCredentials?: boolean;
  credentials?: RequestCredentials;
  keepalive?: boolean;
  priority?: RequestPriority;
  cache?: RequestCache;
  mode?: RequestMode;
  redirect?: RequestRedirect;
  referrer?: string;
  integrity?: string;
  transferCache?: {includeHeaders?: string[]} | boolean;
  timeout?: number;
}



@Injectable({ providedIn: 'root' })
export class UserProfileService {
  private readonly http = inject(HttpClient);
/**
 * @summary Get user profile
 */
 getApiUserProfileUserProfileGet<TData = UserProfileDto>(params: DeepNonNullable<GetApiUserProfileUserProfileGetParams>, options?: HttpClientOptions & { observe?: 'body' }): Observable<TData>;
 getApiUserProfileUserProfileGet<TData = UserProfileDto>(params: DeepNonNullable<GetApiUserProfileUserProfileGetParams>, options?: HttpClientOptions & { observe: 'events' }): Observable<HttpEvent<TData>>;
 getApiUserProfileUserProfileGet<TData = UserProfileDto>(params: DeepNonNullable<GetApiUserProfileUserProfileGetParams>, options?: HttpClientOptions & { observe: 'response' }): Observable<AngularHttpResponse<TData>>;
  getApiUserProfileUserProfileGet<TData = UserProfileDto>(
    params: DeepNonNullable<GetApiUserProfileUserProfileGetParams>, options?: HttpClientOptions & { observe?: any }): Observable<any> {
    return this.http.get<TData>(
      `http://localhost:3000/api/UserProfile/UserProfile_Get`,{
    ...options,
        params: {...params, ...options?.params},}
    );
  }
/**
 * @summary Get current user's profile
 */
 getApiUserProfileUserProfileGetCurrent<TData = UserProfileDto>( options?: HttpClientOptions & { observe?: 'body' }): Observable<TData>;
 getApiUserProfileUserProfileGetCurrent<TData = UserProfileDto>( options?: HttpClientOptions & { observe: 'events' }): Observable<HttpEvent<TData>>;
 getApiUserProfileUserProfileGetCurrent<TData = UserProfileDto>( options?: HttpClientOptions & { observe: 'response' }): Observable<AngularHttpResponse<TData>>;
  getApiUserProfileUserProfileGetCurrent<TData = UserProfileDto>(
     options?: HttpClientOptions & { observe?: any }): Observable<any> {
    return this.http.get<TData>(
      `http://localhost:3000/api/UserProfile/UserProfile_GetCurrent`,options
    );
  }
/**
 * @summary Update current user's profile
 */
 putApiUserProfileUserProfileUpdateCurrentUserProfile<TData = void>(updateUserProfileDto: UpdateUserProfileDto, options?: HttpClientOptions & { observe?: 'body' }): Observable<TData>;
 putApiUserProfileUserProfileUpdateCurrentUserProfile<TData = void>(updateUserProfileDto: UpdateUserProfileDto, options?: HttpClientOptions & { observe: 'events' }): Observable<HttpEvent<TData>>;
 putApiUserProfileUserProfileUpdateCurrentUserProfile<TData = void>(updateUserProfileDto: UpdateUserProfileDto, options?: HttpClientOptions & { observe: 'response' }): Observable<AngularHttpResponse<TData>>;
  putApiUserProfileUserProfileUpdateCurrentUserProfile<TData = void>(
    updateUserProfileDto: UpdateUserProfileDto, options?: HttpClientOptions & { observe?: any }): Observable<any> {
    return this.http.put<TData>(
      `http://localhost:3000/api/UserProfile/UserProfile_UpdateCurrentUserProfile`,
      updateUserProfileDto,options
    );
  }
};