// src/app/core/models/userProfileDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */
import type { UserProfileDtoRolesItem } from './userProfileDtoRolesItem';

export interface UserProfileDto {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  tenantId: string;
  roles: UserProfileDtoRolesItem[];
}