// src/app/core/models/updateUserProfileDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface UpdateUserProfileDto {
  firstName: string;
  lastName: string;
  phoneNumber: string;
  address: string;
  idNumber: string;
  dateOfBirth: string;
}